move files around; gradesModal: autofocus filter element when starting to type
