+++ src/lib/Calendar/CalendarSelector.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
 	import { createEventDispatcher } from 'svelte';
 
-	import TileInteractiveElementWrapper from '$lib/TileInteractiveElementWrapper.svelte';
+	import TileInteractiveElementWrapper from '$lib/TilesAndModals/TileInteractiveElementWrapper.svelte';
 	import { dateIsThisWeek, dateIsToday, getNextWeekday } from '$lib/TSHelpers/DateHelper';
 
 	export let disablePadding = false;
